+++ src/Farmer/Builders/Builders.Storage.fs
@@ -11,11 +11,7 @@ open FileShares
 type StorageAccount =
     /// Gets an ARM Expression connection string for any Storage Account.
     static member getConnectionString (storageAccount:ResourceId) =
-        let expr =
-            sprintf
-                "concat('DefaultEndpointsProtocol=https;AccountName=%s;AccountKey=', listKeys(%s, '2017-10-01').keys[0].value)"
-                storageAccount.Name.Value
-                storageAccount.ArmExpression.Value
+        let expr = $"concat('DefaultEndpointsProtocol=https;AccountName={storageAccount.Name.Value};AccountKey=', listKeys({storageAccount.ArmExpression.Value}, '2017-10-01').keys[0].value)"
         ArmExpression.create(expr, storageAccount)
     /// Gets an ARM Expression connection string for any Storage Account.
     static member getConnectionString (storageAccountName:StorageAccountName, ?group) =
@@ -61,8 +57,8 @@ type StorageAccountConfig =
             .Map(sprintf "%s.primaryEndpoints.web")
     member this.WebsitePrimaryEndpointHost =
         this.WebsitePrimaryEndpoint
-            .Map(fun uri -> sprintf "replace(replace(%s, 'https://', ''), '/', '')" uri)
-    member this.Endpoint = sprintf "%s.blob.core.windows.net" this.Name.ResourceName.Value
+            .Map(fun uri -> $"replace(replace({uri}, 'https://', ''), '/', '')")
+    member this.Endpoint = $"{this.Name.ResourceName.Value}.blob.core.windows.net"
     member this.ResourceId = storageAccounts.resourceId this.Name.ResourceName
     interface IBuilder with
         member this.ResourceId = this.ResourceId
@@ -113,10 +109,7 @@ type StorageAccountConfig =
                 }
             for roleAssignment in this.RoleAssignments do
                 let uniqueName =
-                    sprintf "%s%s%O"
-                        this.Name.ResourceName.Value
-                        roleAssignment.Principal.ArmExpression.Value
-                        roleAssignment.Role.Id
+                    $"{this.Name.ResourceName.Value}{roleAssignment.Principal.ArmExpression.Value}{roleAssignment.Role.Id}"
                     |> DeterministicGuid.create
                     |> string
                     |> ResourceName
@@ -222,7 +215,7 @@ type StorageAccountBuilder() =
                 match state.Sku with
                 | Blobs (replication, _) -> Blobs(replication, Some tier)
                 | GeneralPurpose (V2 (replication, _)) -> GeneralPurpose (V2 (replication, Some tier))
-                | other -> failwithf "You can only set the default access tier for Blobs or General Purpose V2 storage accounts. This account is %A" other
+                | other -> failwith $"You can only set the default access tier for Blobs or General Purpose V2 storage accounts. This account is %A{other}."
         }
     /// Specify network access control lists for this storage account.
     [<CustomOperation "set_network_acls">]