+++ src/Farmer/Builders/Builders.WebApp.fs
@@ -19,6 +19,7 @@ type JavaRuntime =
     member this.Jre = match this with Java8 -> "jre8" | Java11 -> "java11"
 type Runtime =
     | DotNetCore of string
+    | DotNet of version:string
     | Node of string
     | Php of string
     | Ruby of string
@@ -50,6 +51,7 @@ type Runtime =
     static member Java8WildFly14 = Java (Java8, WildFly14)
     static member Java8Tomcat90 = Java (Java8, JavaHost.Tomcat90)
     static member Java8Tomcat85 = Java (Java8, JavaHost.Tomcat85)
+    static member DotNet50 = DotNet "5.0"
     static member AspNet47 = AspNet "4.0"
     static member AspNet35 = AspNet "2.0"
     static member Python27 = Python ("2.7", "2.7")
@@ -62,12 +64,39 @@ module AppSettings =
 
 let publishingPassword (name:ResourceName) =
     let resourceId = config.resourceId (name, ResourceName "publishingCredentials")
-    let expr = sprintf "list(%s, '2014-06-01').properties.publishingPassword" resourceId.ArmExpression.Value
+    let expr = $"list({resourceId.ArmExpression.Value}, '2014-06-01').properties.publishingPassword"
     ArmExpression.create(expr, resourceId)
 
 type SecretStore =
     | AppService
-    | KeyVault of ResourceRef<WebAppConfig>
+    | KeyVault of ResourceRef<CommonWebConfig>
+
+module private WebAppConfig =
+    let ToCommon state =
+        { Name = state.Name
+          ServicePlan = state.ServicePlan
+          AppInsights = state.AppInsights
+          OperatingSystem = state.OperatingSystem
+          Settings = state.Settings
+          Cors = state.Cors
+          Identity = state.Identity
+          SecretStore = state.SecretStore
+          ZipDeployPath = state.ZipDeployPath
+          AlwaysOn = state.AlwaysOn
+          WorkerProcess = state.WorkerProcess }
+    let FromCommon state (config: CommonWebConfig): WebAppConfig =
+        { state with
+            Name = config.Name
+            ServicePlan = config.ServicePlan
+            AppInsights = config.AppInsights
+            OperatingSystem = config.OperatingSystem
+            Settings = config.Settings
+            Cors = config.Cors
+            Identity = config.Identity
+            SecretStore = config.SecretStore
+            ZipDeployPath = config.ZipDeployPath
+            AlwaysOn = config.AlwaysOn
+            WorkerProcess = config.WorkerProcess }
 
 /// Common fields between WebApp and Functions
 type CommonWebConfig =
@@ -78,8 +107,9 @@ type CommonWebConfig =
       Settings : Map<string, Setting>
       Cors : Cors option
       Identity : Identity.ManagedIdentity
+      SecretStore : SecretStore
       ZipDeployPath : string option
-      AlwaysOn : bool 
+      AlwaysOn : bool
       WorkerProcess : Bitness option }
 
 type WebAppConfig =
@@ -121,7 +151,7 @@ type WebAppConfig =
     member this.AppInsightsName = this.AppInsights |> Option.map (fun ai -> ai.resourceId(this.Name).Name)
     /// Gets the ARM expression path to the publishing password of this web app.
     member this.PublishingPassword = publishingPassword (this.Name)
-    member this.Endpoint = sprintf "%s.azurewebsites.net" this.Name.Value
+    member this.Endpoint = $"{this.Name.Value}.azurewebsites.net"
     member this.SystemIdentity = SystemIdentity this.ResourceId
     member this.ResourceId = sites.resourceId this.Name
     interface IBuilder with
@@ -129,7 +159,7 @@ type WebAppConfig =
         member this.BuildResources location = [
             let keyVault, secrets =
                 match this.SecretStore with
-                | KeyVault (DeployableResource this vaultName) ->
+                | KeyVault (DeployableResource (WebAppConfig.ToCommon this) vaultName) ->
                     let store = keyVault {
                         name vaultName.Name
                         add_access_policy (AccessPolicy.create (this.SystemIdentity.PrincipalId, [ KeyVault.Secret.Get ]))
@@ -212,7 +242,7 @@ type WebAppConfig =
                     match this.DockerAcrCredentials with
                     | Some credentials ->
                         "DOCKER_REGISTRY_SERVER_PASSWORD", ParameterSetting credentials.Password
-                        Setting.AsLiteral ("DOCKER_REGISTRY_SERVER_URL", sprintf "https://%s.azurecr.io" credentials.RegistryName)
+                        Setting.AsLiteral ("DOCKER_REGISTRY_SERVER_URL", $"https://{credentials.RegistryName}.azurecr.io")
                         Setting.AsLiteral ("DOCKER_REGISTRY_SERVER_USERNAME", credentials.RegistryName)
                     | None ->
                         ()
@@ -225,14 +255,14 @@ type WebAppConfig =
                      | AppService ->
                          this.Settings
                      | KeyVault r ->
-                        let name = r.resourceId this
+                        let name = r.resourceId (WebAppConfig.ToCommon this)
                         [ for setting in this.Settings do
                             match setting.Value with
                             | LiteralSetting _ ->
                                 setting.Key, setting.Value
                             | ParameterSetting _
                             | ExpressionSetting _ ->
-                                setting.Key, LiteralSetting (sprintf "@Microsoft.KeyVault(SecretUri=https://%s.vault.azure.net/secrets/%s)" name.Name.Value setting.Key)
+                                setting.Key, LiteralSetting $"@Microsoft.KeyVault(SecretUri=https://{name.Name.Value}.vault.azure.net/secrets/{setting.Key})"
                         ] |> Map.ofList
                     ) |> Map.toList)
                 |> Map
@@ -275,19 +305,22 @@ type WebAppConfig =
                         Some ("DOCKER|" + image)
                     | None ->
                         match this.Runtime with
-                        | DotNetCore version -> Some ("DOTNETCORE|" + version)
-                        | Node version -> Some ("NODE|" + version)
-                        | Php version -> Some ("PHP|" + version)
-                        | Ruby version -> Some ("RUBY|" + version)
-                        | Java (runtime, JavaSE) -> Some (sprintf "JAVA|%d-%s" runtime.Version runtime.Jre)
-                        | Java (runtime, (Tomcat version)) -> Some (sprintf "TOMCAT|%s-%s" version runtime.Jre)
-                        | Java (Java8, WildFly14) -> Some (sprintf "WILDFLY|14-%s" Java8.Jre)
-                        | Python (linuxVersion, _) -> Some (sprintf "PYTHON|%s" linuxVersion)
+                        | DotNetCore version -> Some $"DOTNETCORE|{version}"
+                        | Node version -> Some $"NODE|{version}"
+                        | Php version -> Some $"PHP|{version}"
+                        | Ruby version -> Some $"RUBY|{version}"
+                        | Java (runtime, JavaSE) -> Some $"JAVA|{runtime.Version}-{runtime.Jre}"
+                        | Java (runtime, (Tomcat version)) -> Some $"TOMCAT|{version}-{runtime.Jre}"
+                        | Java (Java8, WildFly14) -> Some $"WILDFLY|14-{Java8.Jre}"
+                        | Python (linuxVersion, _) -> Some $"PYTHON|{linuxVersion}"
                         | _ -> None
               NetFrameworkVersion =
                 match this.Runtime with
-                | AspNet version -> Some (sprintf "v%s" version)
-                | _ -> None
+                | AspNet version
+                | DotNet ("5.0" as version) ->
+                    Some $"v{version}"
+                | _ ->
+                    None
               JavaVersion =
                 match this.Runtime, this.OperatingSystem with
                 | Java (Java11, Tomcat _), Windows -> Some "11"
@@ -315,7 +348,7 @@ type WebAppConfig =
                 | Php _, _ -> Some "php"
                 | Python _, Windows -> Some "python"
                 | DotNetCore _, Windows -> Some "dotnetcore"
-                | AspNet _, _ -> Some "dotnet"
+                | AspNet _, _ | DotNet "5.0", Windows -> Some "dotnet"
                 | _ -> None
                 |> Option.map(fun stack -> "CURRENT_STACK", stack)
                 |> Option.toList
@@ -409,14 +442,17 @@ type WebAppBuilder() =
         { state with
             SiteExtensions =
                 match state with
-                | { Runtime = Runtime.DotNetCore _; AutomaticLoggingExtension = true } ->
+                // its important to only add this extension if we're not using Web App for Containers - if we are
+                // then this will generate an error during deployment:
+                // No route registered for '/api/siteextensions/Microsoft.AspNetCore.AzureAppServices.SiteExtension'
+                | { Runtime = Runtime.DotNetCore _; AutomaticLoggingExtension = true ; DockerImage = None } ->
                     state.SiteExtensions.Add WebApp.Extensions.Logging
                 | _ ->
                     state.SiteExtensions
             DockerImage =
                 match state.DockerImage, state.DockerAcrCredentials with
                 | Some (image, tag), Some credentials when not (image.Contains "azurecr.io") ->
-                    Some (sprintf "%s.azurecr.io/%s" credentials.RegistryName image, tag)
+                    Some ($"{credentials.RegistryName}.azurecr.io/{image}", tag)
                 | Some x, _ ->
                     Some x
                 | None, _ ->
@@ -479,7 +515,7 @@ type WebAppBuilder() =
         { state with
             DockerAcrCredentials =
                 Some {| RegistryName = registryName
-                        Password = SecureParameter (sprintf "docker-password-for-%s" registryName) |} }
+                        Password = SecureParameter $"docker-password-for-{registryName}" |} }
     [<CustomOperation "source_control">]
     member _.SourceControl(state:WebAppConfig, url, branch) =
         { state with
@@ -496,24 +532,6 @@ type WebAppBuilder() =
     member this.EnableCi(state:WebAppConfig) = this.SourceControlCi(state, Enabled)
     [<CustomOperation "disable_source_control_ci">]
     member this.DisableCi(state:WebAppConfig) = this.SourceControlCi(state, Disabled)
-    /// Creates a key vault instance. All secret settings will automatically be mapped into key vault.
-    [<CustomOperation "use_keyvault">]
-    member _.UseKeyVault (state:WebAppConfig) =
-        { state with
-            Identity = { state.Identity with SystemAssigned = Enabled }
-            SecretStore = KeyVault (derived(fun c -> vaults.resourceId (ResourceName (c.Name.Value + "vault")))) }
-    /// Links your application to a Farmer-managed key vault instance. All secret settings will automatically be mapped into key vault.
-    [<CustomOperation "link_to_keyvault">]
-    member _.LinkToKeyVault (state:WebAppConfig, vaultName:ResourceName) =
-        { state with
-            Identity = { state.Identity with SystemAssigned = Enabled }
-            SecretStore = KeyVault (External(Managed (vaults.resourceId vaultName))) }
-    /// Links your application to an existing key vault instance. All secret settings will automatically be mapped into key vault.
-    [<CustomOperation "link_to_unmanaged_keyvault">]
-    member _.LinkToExternalKeyVault(state:WebAppConfig, resourceId) =
-        { state with
-            Identity = { state.Identity with SystemAssigned = Enabled }
-            SecretStore = KeyVault (External(Unmanaged resourceId)) }
     [<CustomOperation "add_extension">]
     member _.AddExtension (state:WebAppConfig, extension) = { state with SiteExtensions = state.SiteExtensions.Add extension }
     member this.AddExtension (state:WebAppConfig, name) = this.AddExtension (state, ExtensionName name)
@@ -523,29 +541,8 @@ type WebAppBuilder() =
     interface ITaggable<WebAppConfig> with member _.Add state tags = { state with Tags = state.Tags |> Map.merge tags }
     interface IDependable<WebAppConfig> with member _.Add state newDeps = { state with Dependencies = state.Dependencies + newDeps }
     interface IServicePlanApp<WebAppConfig> with
-        member _.Get state =
-            { Name = state.Name
-              ServicePlan = state.ServicePlan
-              AppInsights = state.AppInsights
-              OperatingSystem = state.OperatingSystem
-              Settings = state.Settings
-              Cors = state.Cors
-              Identity = state.Identity
-              ZipDeployPath = state.ZipDeployPath
-              AlwaysOn = state.AlwaysOn 
-              WorkerProcess = state.WorkerProcess }
-        member _.Wrap state config =
-            { state with
-                Name = config.Name
-                ServicePlan = config.ServicePlan
-                AppInsights = config.AppInsights
-                OperatingSystem = config.OperatingSystem
-                Settings = config.Settings
-                Cors = config.Cors
-                Identity = config.Identity
-                ZipDeployPath = config.ZipDeployPath
-                AlwaysOn = config.AlwaysOn 
-                WorkerProcess = config.WorkerProcess }
+        member _.Get state = WebAppConfig.ToCommon state
+        member _.Wrap state config = WebAppConfig.FromCommon state config
 
 let webApp = WebAppBuilder()
 
@@ -582,7 +579,7 @@ module Extensions =
         /// Instead of creating a new service plan instance, configure this webapp to point to another unmanaged service plan instance.
         /// A dependency will automatically be set for this instance.
         [<CustomOperation "link_to_unmanaged_service_plan">]
-        member this.LinkToUnmanagedServicePlan (state:'T, resourceId) = { this.Get state with ServicePlan = External (Unmanaged resourceId) } |> this.Wrap state
+        member this.LinkToUnmanagedServicePlan (state:'T, resourceId) = { this.Get state with ServicePlan = unmanaged resourceId } |> this.Wrap state
         /// Sets the name of the automatically-created app insights instance.
         [<CustomOperation "app_insights_name">]
         member this.UseAppInsights (state:'T, name) = { this.Get state with AppInsights = Some (named components name) } |> this.Wrap state
@@ -600,7 +597,7 @@ module Extensions =
         /// Instead of creating a new AI instance, configure this webapp to point to an unmanaged AI instance.
         /// A dependency will not be set for this instance.
         [<CustomOperation "link_to_unmanaged_app_insights">]
-        member this.LinkUnmanagedAppInsights (state:'T, resourceId) = { this.Get state with AppInsights = Some (External(Unmanaged resourceId)) } |> this.Wrap state
+        member this.LinkUnmanagedAppInsights (state:'T, resourceId) = { this.Get state with AppInsights = Some (unmanaged resourceId) } |> this.Wrap state
         /// Sets an app setting of the web app in the form "key" "value".
         [<CustomOperation "setting">]
         member this.AddSetting (state:'T, key, value) =
@@ -676,3 +673,27 @@ module Extensions =
         ///Chooses the bitness (32 or 64) of the worker process
         [<CustomOperation "worker_process">]
         member this.WorkerProcess (state:'T, bitness) = { this.Get state with WorkerProcess = Some bitness } |> this.Wrap state
+        /// Creates a key vault instance. All secret settings will automatically be mapped into key vault.
+        [<CustomOperation "use_keyvault">]
+        member this.UseKeyVault (state:'T) =
+            let current = this.Get state
+            { current with
+                Identity = { current.Identity with SystemAssigned = Enabled }
+                SecretStore = KeyVault (derived(fun c -> vaults.resourceId (ResourceName (c.Name.Value + "vault")))) }
+            |> this.Wrap state
+        /// Links your application to a Farmer-managed key vault instance. All secret settings will automatically be mapped into key vault.
+        [<CustomOperation "link_to_keyvault">]
+        member this.LinkToKeyVault (state:'T, vaultName:ResourceName) =
+            let current = this.Get state
+            { current with
+                Identity = { current.Identity with SystemAssigned = Enabled }
+                SecretStore = KeyVault (managed vaults vaultName) }
+            |> this.Wrap state
+        /// Links your application to an existing key vault instance. All secret settings will automatically be mapped into key vault.
+        [<CustomOperation "link_to_unmanaged_keyvault">]
+        member this.LinkToExternalKeyVault(state:'T, resourceId) =
+            let current = this.Get state
+            { current with
+                Identity = { current.Identity with SystemAssigned = Enabled }
+                SecretStore = KeyVault (unmanaged resourceId) }
+            |> this.Wrap state