+++ src/Tests/AllTests.fs
@@ -4,6 +4,8 @@ open Expecto
 open System
 open Farmer
 
+let hasEnv a b = Environment.GetEnvironmentVariable a = b
+
 [<Tests>]
 let allTests =
     testSequencedGroup "" <|
@@ -41,13 +43,14 @@ let allTests =
                 VirtualNetworkGateway.tests
                 Databricks.tests
                 JsonRegression.tests
+                Types.tests
                 AzCli.tests
             ]
             testList "Control" [
                 Template.tests
                 Identity.tests
                 Common.tests
-                if Environment.GetEnvironmentVariable "TF_BUILD" = "True" then AzCli.endToEndTests
+                if hasEnv "TF_BUILD" "True" || hasEnv "FARMER_E2E" "True" then AzCli.endToEndTests
             ]
         ]