+++ src/Tests/JsonRegression.fs
@@ -17,7 +17,7 @@ let tests =
             let path = __SOURCE_DIRECTORY__ + "/test-data/" + jsonFile
             let expected = File.ReadAllText path
             let actual = template.Template |> Writer.toJson
-            Expect.equal expected actual (sprintf "ARM template generation has changed! Either fix the writer, or update the contents of the generated file (%s)" path)
+            Expect.equal (actual.Trim()) (expected.Trim()) (sprintf "ARM template generation has changed! Either fix the writer, or update the contents of the generated file (%s)" path)
 
         test "Generates lots of resources" {
             let number = string 1979
@@ -44,9 +44,17 @@ let tests =
                     }
                 ]
             }
+            let cosmosMongo = cosmosDb {
+                name "testdbmongo"
+                account_name "testaccountmongo"
+                kind Mongo
+                throughput 400<CosmosDb.RU>
+                failover_policy CosmosDb.NoFailover
+                consistency_policy (CosmosDb.BoundedStaleness(500, 1000))
+            }
 
             compareResourcesToJson
-                [ sql; storage; web; fns; svcBus; cdn; containerGroup; cosmos ]
+                [ sql; storage; web; fns; svcBus; cdn; containerGroup; cosmos; cosmosMongo ]
                 "lots-of-resources.json"
         }
 
@@ -88,5 +96,51 @@ let tests =
                 ]
             }
             compareResourcesToJson [ data; web; hub; logs; mydiagnosticSetting ] "diagnostics.json"
+        }
+
+        test "Event Grid" {
+            let storageSource = storageAccount {
+                name "isaacgriddevprac"
+                add_private_container "data"
+                add_queue "todo"
+            }
+
+            let eventQueue = queue { name "events" }
+            let sb = serviceBus {
+                name "farmereventpubservicebusns"
+                add_queues [ eventQueue ]
+            }
+
+            let eventHubGrid = eventGrid {
+                topic_name "newblobscreated"
+                source storageSource
+                add_queue_subscriber storageSource "todo" [ SystemEvents.Storage.BlobCreated ]
+                add_servicebus_queue_subscriber sb eventQueue [ SystemEvents.Storage.BlobCreated ]
+            }
+
+            compareResourcesToJson [ storageSource; sb; eventHubGrid ] "event-grid.json"
+        }
+
+        test "Can parse JSON into an ARM template" {
+            let json = """    {
+      "apiVersion": "2019-06-01",
+      "dependsOn": [],
+      "kind": "StorageV2",
+      "location": "northeurope",
+      "name": "jsontest",
+      "properties": {},
+      "sku": {
+        "name": "Standard_LRS"
+      },
+      "tags": {},
+      "type": "Microsoft.Storage/storageAccounts"
+    }
+"""
+            let resource = arm { add_resource (Resource.ofJson json) } |> Storage.getStorageResource
+            printfn "%A" resource
+
+            Expect.equal resource.Name "jsontest" "Account name is wrong"
+            Expect.equal resource.Sku.Name "Standard_LRS" "SKU is wrong"
+            Expect.equal resource.Kind "StorageV2" "Kind"
         }
     ]