+++ RELEASE_NOTES.md
@@ -1,13 +1,29 @@
 Release Notes
 =============
-## vNext
+
+##  1.5.0-beta
 * Container Groups: Support for init containers.
-* Framework: Add support for NET5, upgrade to F#5
+* Container Groups: Support for liveliness and readiness probes on containers.
+* Container Groups: Connect network profile to an existing virtual network.
+* Container Groups: Bugfix for outputs.
+* CosmosDb: Add support for MongoDB as a database kind.
+* Event Grid: Ensure destination Queues are created as a dependency
+* Event Grid: Add ServiceBus Queue and Topic as supported destinations
 * Functions: Support for 64 bits.
+* Functions: Add option to use managed Key Vault
+* KeyVault: Fix an issue with adding tags on main KeyVault builder.
+* ServiceBus: update namespace validation rules to follow [Microsoft documentation](https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftservicebus)
 * Storage: Add support for tables
 * Storage: Support for firewall to restrict storage account network access to virtual network subnets, IP addresses, and CIDR prefixes.
 * Virtual Network: Support for creating service endpoints on subnets.
 * Virtual Network: Support for assigning existing service endpoint policies to subnets.
+* Web App: Disables the automatic addition of the logging site extension when `docker_image` is used
+* Web App: Add dotnet 5.0 runtime option
+
+* Framework: Updated DeterministicGuid for RFC 4122 compatibility
+* Framework: Add support for NET5, upgrade to F#5.
+* Framework: Simplify Event Grid builder
+* Framework: Use System.Text.Json instead of Newtonsoft.Json
 
 ## 1.4.0
 * Bing Search: Support for Bing Search (migrated from Cognitive Services).